@@ -2,11 +2,12 @@
 
 namespace App\Events;
 
+use App\Models\Collection;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Queue\SerializesModels;
 
-class FollowedEvent
+class SubscribedEvent
 {
     use SerializesModels;
 
@@ -25,11 +26,11 @@ class FollowedEvent
 
     public function handle()
     {
-        if ($this->model instanceof User) {
+        if ($this->model instanceof Collection) {
             // 接受通知的用户
-            $this->to = $this->model;
+            $this->to = $this->model->user;
             $this->message = [
-                'message' => $this->from->username . '关注了你 (＾▽＾)ｺ 快去看看吧'
+                'message' => $this->from->username . '订阅了你的收藏集: ' . subtext($this->model->title, 8) . ' (＾▽＾)ｺ 快去看看吧'
             ];
         }
     }
